+++ NEWS.md
@@ -1,6 +1,9 @@
 affiner 2.0.0 (development)
 ===========================
 
+  + `is_equivalent()` is a generic S3 method which tests whether two R objects are "equivalent".
+    The `is_equivalent()` method for angle vectors tests whether two angles are congruent.
+    The `is_equivalent()` method for `Point1D`, `Line2D`, `Plane3D` classes tests whether they are the same point/line/plane after standardization.
 * If R >= 4.4.0 can now also use `crossprod()` (in addition to `cross_product3d()`) to compute the
   **vector** cross product of two `Coord3D` objects.
 * Can now use `ceiling()`, `floor()`, `round()`, `signif()`, and `trunc()` on `Coord1D`, `Coord2D`, and `Coord3D` objects (#63).